+++ .gitignore
@@ -25,3 +25,4 @@ dist-ssr
 
 # env
 .env
+.cypress.env.json
