+++ src/api/getMyPlaylists.ts
@@ -1,5 +1,5 @@
 import axiosInstance from "@/api/axiosInstance";
-import { PlaylistCardData } from "@/api/playlistCardData";
+import { PlaylistCardData } from "@/db/playlistCardData";
 
 // 내가 만든 플레이리스트
 export async function getMyPlaylists(
