+++ src/api/useFollowCount.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getFollowerCount, getFollowingCount } from "@/api/follow";
+import { getFollowerCount, getFollowingCount } from "@/db/follow";
 
 const useFollowCount = (targetId?: number) => {
   const { data: followerCount = 0, isLoading: isFollowerLoading } = useQuery({
