+++ src/api/useOtherUser.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import { getUserByRandomId } from "@/api/users";
-import { User } from "@/api/users";
+import { getUserByRandomId } from "@/db/users";
+import { User } from "@/db/users";
 
 const useOtherUser = (randomId: number) => {
   return useQuery<User | null>({
