+++ src/db/getPlaylistDetail.ts
