+++ src/pages/followInfo/FollowInfo.tsx
@@ -4,8 +4,8 @@ import Title from "@/shared/component/Title";
 import CommonInput from "@/shared/component/input";
 import styled from "@emotion/styled";
 import defaultProfile from "@/assets/images/defaultProfile.svg";
-import useFollowList from "@/api/services/useFollowList";
-import { getUser } from "@/api/users.ts";
+import useFollowList from "@/api/useFollowList";
+import { getUser } from "@/db/users";
 import { useQuery } from "@tanstack/react-query";
 import Loading from "@/shared/component/Loading";
 import Reset from "@/assets/images/reset.svg";
