+++ src/pages/profile/hooks/useUpdateUserInfo.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { updateUser } from "@/api/users";
+import { updateUser } from "@/db/users";
 import { useUserStore } from "@/stores/userStore";
 
 interface UpdateUserInfoParams {
